// backend/src/modules/auth/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'admin',
  SUPER_ADMIN = 'super_admin',
  RESTAURANT_ADMIN = 'restaurant_admin',
  RESTAURANT_OWNER = 'restaurant_owner',
  RESTAURANT_STAFF = 'restaurant_staff',
  DELIVERY_PERSON = 'delivery_person',
  DELIVERY_DRIVER = 'delivery_driver',
  CUSTOMER = 'customer',
}